vnet-flow-logs/requirements.txt
azure-functions==1.23.0
requests>=2.32.3
requests-toolbelt==1.0.0
azure-storage-blob==12.28.0
